+++ src/main.zig
@@ -115,7 +115,7 @@ fn milliTimestamp() i64 {
 
 const ROOT = paths.ROOT;
 const PREFIX = paths.PREFIX;
-const VERSION = "0.1.198";
+const VERSION = "0.1.199";
 
 pub fn main(init: std.process.Init) !void {
     g_io = init.io;